+++ include/viua/bytecode/opcodes.h
@@ -128,13 +128,12 @@ enum OPCODE : viua::internals::types::byte {
     LINK,       // dynamically link native library
 
     CLASS,      // create a prototype for new class
-    PROTOTYPE,  // create a prototype from existing class
     DERIVE,     // derive a prototype from an existing class
     ATTACH,     // attach a method to the prototype
     REGISTER,   // register a prototype in VM's typesystem
 
     NEW,        // construct new instance of a class in a register
-    MSG,        // send a message to an object (used for dynamic dispatch, for static use plan "CALL")
+    MSG,        // send a message to an object (used for dynamic dispatch, for static use plain "CALL")
     INSERT,     // insert an object as a value of an attribute of another object
     REMOVE,     // remove an attribute from an object